Parser

Documentation

parseAst

Arguments

:: String
The code to parse

-> Maybe StatementNode
The parsed AST (or Nothing if parsing fails)

Parses the input code and returns a Just value containing the parsed AST if successful, or Nothing otherwise.

This function parses a series of statements and returns the top-level block node.

Example:

>>> parseAst "var x = 5; print x;"
Just (BlockNode [VariableDeclarationNode "x" (NumberNode 5.0),PrintStatementNode (IdentNode "x")])

data StatementNode

Represents various nodes in a statement.

Constructors

BlockNode [StatementNode]
Block node, consisting of a list of statements.

WhileNode ExpressionNode StatementNode
While loop node, consisting of a loop condition expression and a loop body statement.

IfNode ExpressionNode StatementNode StatementNode
If statement node, consisting of a condition expression, a consequence statement, and an optional alternative statement.

VariableDeclarationNode String ExpressionNode
Variable declaration node, representing the declaration of a variable with an initial value expression.

VariableAssignNode String ExpressionNode
Variable assignment node, representing the assignment of a value expression to a variable.

PrintStatementNode ExpressionNode
Print statement node, representing the output of an expression.

ArrayAssignNode ExpressionNode ExpressionNode
Array assignment node, representing the assignment of a value expression to an element in an array.

data ExpressionNode

Represents various nodes in an expression.

Constructors

BinaryExpressionNode Operator ExpressionNode ExpressionNode
Binary expression node, consisting of an operator and two sub-expressions.

UnaryExpressionNode Operator ExpressionNode
Unary expression node, consisting of an operator and a sub-expression.

StringNode String
String literal node.

BooleanNode Bool
Boolean literal node.

IdentNode String
Identifier node.

NilNode
Nil literal node.

NumberNode Float
Numeric literal node.

ArrayNode [ExpressionNode]
Array literal node, consisting of a list of expressions.

ArrayAccessNode ExpressionNode ExpressionNode
Array access node, representing accessing an element in an array using an index expression.

data Operator

Represents various operators used in expressions.

Constructors

Add
Addition operator (+)

Subtract
Subtraction operator (-)

Multiply
Multiplication operator (*)

Divide
Division operator (/)

Modulo
Modulo operator (%)

Lt
Less than operator (<)

Lte
Less than or equal to operator (<=)

Gt
Greater than operator (>)

Gte
Greater than or equal to operator (>=)

Eq
Equality operator (==)

NotEq
Inequality operator (/=)

expression

Arguments

:: Parser ExpressionNode
The parsed expression

Parses an expression.

This parser is the entry point for parsing expressions.

Example:

>>> parse expression "x + 10 * (y - 5)"
[(BinaryExpressionNode Add (IdentNode "x") (BinaryExpressionNode Multiply (NumberNode 10.0) (BinaryExpressionNode Subtract (IdentNode "y") (NumberNode 5.0))),"")]